create authorization to delete or edit campground

+++ routes/campgrounds.js
@@ -38,11 +38,20 @@ Router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res)=> {
         req.flash('error', 'Cannot find that campground!')
         return res.redirect('/campgrounds')
     }
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error', "You don't have permission to do that!")
+        return res.redirect(`/campgrounds/${id}`)
+    }
     res.render('campgrounds/edit', {campground})
 }))
 Router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res)=> {
     const id = req.params.id
-    const campground = await Campground.findByIdAndUpdate(id, {...req.body.campground})
+    const campground = await Campground.findById(id)
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error', "You don't have permission to do that!")
+        return res.redirect(`/campgrounds/${id}`)
+    }
+    const camp = await Campground.findByIdAndUpdate(id, {...req.body.campground})
     req.flash('success', 'Successfully updated campground!')
     res.redirect(`/campgrounds/${campground._id}`)
 }))
@@ -55,6 +64,11 @@ Router.post('/', isLoggedIn, validateCampground, catchAsync(async (req, res, nex
 }))
 Router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
     const id = req.params.id
+    const campground = await Campground.findById(id)
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error', "You don't have permission to do that!")
+        return res.redirect(`/campgrounds/${id}`)
+    }
     await Campground.findByIdAndDelete(id)
     req.flash('success', 'Successfully deleted campground!')
     res.redirect('/campgrounds')
